Project folder restructure.
