+++ lib/src/features/authentication/applications/auth_page.dart
@@ -1,7 +1,7 @@
-import 'package:fitgap/pages/login_register_switch.dart';
+import 'package:fitgap/src/features/authentication/applications/login_signup_switch.dart';
 import 'package:flutter/material.dart';
-import 'package:fitgap/services/auth.dart';
-import 'package:fitgap/pages/home.dart';
+import 'package:fitgap/src/utils/auth/auth.dart';
+import 'package:fitgap/src/features/home/home.dart';
 // import 'package:fitgap/pages/login.dart';
 
 //Main Page : Auth/Home
