+++ lib/src/features/authentication/models/auth_button.dart
